refactor maas request job request class into module
