+++ python/lib/communication/dmod/communication/maas_request.py
@@ -12,41 +12,6 @@ from numbers import Number
 from typing import Dict, List, Optional, Set, Union
 
 
-class DmodJobRequest(AbstractInitRequest, ABC):
-    """
-    The base class underlying all types of messages requesting execution of some kind of workflow job.
-    """
-
-    def __int__(self, *args, **kwargs):
-        super(DmodJobRequest, self).__int__(*args, **kwargs)
-
-    @property
-    @abstractmethod
-    def data_requirements(self) -> List[DataRequirement]:
-        """
-        List of all the explicit and implied data requirements for this request, as needed for creating a job object.
-
-        Returns
-        -------
-        List[DataRequirement]
-            List of all the explicit and implied data requirements for this request.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def output_formats(self) -> List[DataFormat]:
-        """
-        List of the formats of each required output dataset for the requested job.
-
-        Returns
-        -------
-        List[DataFormat]
-            List of the formats of each required output dataset for the requested job.
-        """
-        pass
-
-
 class ModelExecRequest(ExternalRequest, DmodJobRequest, ABC):
     """
     An abstract extension of ::class:`DmodJobRequest` for requesting model execution jobs.
